requirements.txt
# Core Dependencies
streamlit>=1.28.0
# AI/ML (Optional - for fine-tuned models)
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0

# Google Calendar Integration (Optional)
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# PDF Processing (Optional)
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Document Processing (Optional)
python-docx>=0.8.11

# Utilities
python-dateutil>=2.8.2